+++ lasso/sub_fft.h
@@ -7,7 +7,7 @@
 #include "mir_hist1d_ope.h"
 
 void GetProxMapFft(const double* const h_arr, long nrow,
-                   const int* const index_arr, // nrow
+                   const long* const index_arr, // nrow
                    const double* const y_arr, long ncol,
                    double delta_freq,
                    double lconst,
@@ -15,7 +15,7 @@ void GetProxMapFft(const double* const h_arr, long nrow,
                    double* const pL_arr);  // ncol
 
 double QFuncFft(const double* const h_arr, long nrow,
-                const int* const index_arr, // nrow
+                const long* const index_arr, // nrow
                 const double* const x_arr, long ncol,
                 const double* const y_arr, // ncol
                 double delta_freq,
@@ -23,26 +23,26 @@ double QFuncFft(const double* const h_arr, long nrow,
                 double lambda);
 
 void DiffFFuncFft(const double* const h_arr, long nrow,
-                  const int* const index_arr, // nrow
+                  const long* const index_arr, // nrow
                   const double* const x_arr, long ncol,
                   double delta_freq,
                   double* const out_arr); // ncol
 
 double FFuncFft(const double* const h_arr, long nrow,
-                const int* const index_arr, // nrow
+                const long* const index_arr, // nrow
                 const double* const x_arr, long ncol,
                 double delta_freq);
 
 double GFuncFft(const double* const x_arr, long ncol, double lambda);
 
 double FGFuncFft(const double* const h_arr, long nrow,
-                 const int* const index_arr, // nrow
+                 const long* const index_arr, // nrow
                  const double* const x_arr, long ncol,
                  double delta_freq,
                  double lambda);
 
 double GetLconstFft(const double* const h_arr, long nrow,
-                    const int* const index_arr, // nrow
+                    const long* const index_arr, // nrow
                     const double* const x_arr, long ncol,
                     double delta_freq,
                     double lconst,
@@ -50,12 +50,12 @@ double GetLconstFft(const double* const h_arr, long nrow,
                     double lambda);
 
 void AxFft(const double* const x_arr, long ncol, 
-           const int* const index_arr, long nrow,
+           const long* const index_arr, long nrow,
            double delta_freq,
            double* const out_arr); // nrow
 
 void ATyFft(const double* const y_arr, long nrow, 
-            const int* const index_arr, // nrow
+            const long* const index_arr, // nrow
             double delta_freq,
             long ncol,
             double* const out_arr); // ncol
